@@ -1,7 +1,7 @@
 import logging
 
 # from tests.common.log_glue import *
-from tests.common.log_glue_incl import KEY_LOG_GLUE, TEST_CONTEXT, ret_dict_info
+from tests.common.log_glue_incl import KEY_LOG_GLUE, TEST_CONTEXT, old_ret_dict_info
 
 from pytest_bdd import parsers, given, when, then   # isort:skip
 
